Experiment: add array repetition to rust_val
Would nicely allow representing stuff like `[(); usize::MAX]`, or in general very large arrays made from repeated values.
However, it currently is super clumsy and breaks a lot of stuff due to not always breaking the array, and its not easy to know when we should or shouldnt break those...

For now this isnt worth it, I think. Maybe once i have redone the encoding system to be more flexible it will work better, in particular with proper layouts. What I currently have is really not great : (